refactor(nvim): clueをinit.luaに移動
全カテゴリをまたぐキーバインドヒントなのでエントリポイントに配置
